chore: 타입 스크립트 초기 설정
