+++ src/index.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import App from "./App";
+import { $ } from "utils/selector";
+
+const root = ReactDOM.createRoot($("root"));
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
